mergeSTR runner creation
